feat(bgl): the vertex-pulling arm of the forward shader (W3) (#131)
* feat(bgl): the vertex-pulling arm of the forward shader (W3)

Forward_StaticMesh gains a BGL_WGSL VSMain that walks the same meshlet ->
vertexMap -> byte-decode chain MSMain walks, addressed off SV_VertexID: the
expansion pass writes one MeshletInstance record per (instance, meshlet), and the
draw covers each with cMaxPrimsPerMeshlet * 3 vertices. Primitives past the
meshlet's triangleCount collapse to a point rather than a NaN position, which WGSL
will not const-fold. The amplification and mesh entries are guarded out, as Slang
validates every entry in a module against the target at load.

Getting there needed four portability fixes, each found by compiling the real
shader rather than by inspection:

- Meshlet's uint16_t counts and PsoType's uint16_t underlying type: core WGSL has
  no 16-bit scalar. Both are widened. Meshlet is CPU-built by Scene at load, not
  read from the cooked format, so nothing constrains its width -- the doc claimed
  otherwise and is corrected. A 16-bit enum is the easier one to miss: it fails
  where the enum is *used*, as an empty parameter type.
- Mesh packed to 72 bytes on the CPU but has an 80-byte WGSL stride, since WGSL
  rounds a struct's size up to its alignment and the float4x4 makes that 16. Padded
  in the IDL source so both mirrors agree; the second element would otherwise be
  read from the wrong offset.
- ForwardData's bucket bases and compaction output are read-write, which a WebGPU
  vertex stage may not bind. They exist for the amplification stage, so they are
  guarded to it, and the records carry an already-resolved instance index. A
  read-write slot now drops Vertex from its visibility, so a buffer some other
  entry point declares cannot invalidate the whole layout.
- The debug buffer is an atomic append and so read-write too. Asserts in code a
  vertex stage can reach are gated on BGL_DBG_VERTEX_SAFE.

Verified on Metal/Dawn against hand-built geometry -- one submesh, one meshlet, one
triangle, NDC positions under an identity transform -- so a wrong link anywhere in
the chain moves or drops the triangle.


* fix(bgl/idl): spell the PsoType casts uint32_t, not uint

idlgen copies enum and constant initializers across verbatim, so a cast in one has
to be valid C++ as well as Slang. `uint` is not a C++ type, and the generated
PsoType.h failed both MSVC legs -- everything else in those builds cascaded from
it. The uint16_t spelling this replaced happened to be valid in both, which is why
the constraint went unnoticed until the widening.

A local macOS build does not catch this: sys/types.h defines `uint` as a POSIX
typedef, so clang accepts the generated header that MSVC rejects.


* docs(idlgen): an IDL initializer must be valid C++ as well as Slang

The verbatim copy is documented; the constraint it implies was not, and `uint`
passes a macOS build because sys/types.h provides it as a POSIX typedef.


* refactor(bgl/shaders): move ASBase to its own module, drop two ifdef arms

Review follow-up on #131.

ASBase reads the fields ForwardData carries only for the amplification stage, so
it moves to forward/amplification.slang, which only the mesh-shader arm imports --
the WGSL target never loads it and common.slang needs no guard at all.

meshletInstances goes back among the unconditional fields, leaving ForwardData with
one `#if !defined(BGL_WGSL)` instead of an if/else. The mesh path gets the same
records as a payload and leaves the buffer unread.

Drops the two comments that restated what the widened types already say; the
constraints behind them live in docs/slang_shaders.md and docs/idlgen.md.


* test(bgl): pin every IDL struct's stride against both GPU targets

A struct a shader loads must have one stride on the CPU and on the GPU, or element
N is read from the wrong offset -- silently, since nothing round-trips a second
element. Dawn caught the Mesh case only by accident, on a single-element buffer
whose minBindingSize it happened to check.

Reflecting the same probe module for WGSL and DXIL side by side shows the shape of
the problem: DXIL agrees with the CPU mirror everywhere, and WGSL is the outlier,
because it alone rounds a struct's size up to its alignment. So a fix has to change
the shared Slang struct -- padding either mirror alone would leave the other target
disagreeing. That is why Mesh's pad, which lives in the IDL source, lands all three
at 80.

Two structs are divergent and listed rather than fixed: PbrMaterial (52 vs 64) and
LoosePbrMaterial (136 vs 176). Nothing portable loads them yet -- the material atlas
is W4 -- and padding them changes the material upload path, which belongs there.


* refactor(bgl/shaders): rename the debug gate, drop the module comment

Review follow-up on #131.

BGL_DBG_VERTEX_SAFE read as a property of the vertex stage rather than of the debug
facility. BERNINI_GPU_DEBUG_ALL_STAGES says what the condition actually is -- GPU
debug is on *and* usable from every stage -- and shares the prefix of the macro it
narrows, so a site reads as "assert only where an assert can run".


* feat(bgl/idl): idlgen pads every struct to its WGSL stride

WGSL alone rounds a struct's size up to its alignment, so a struct can have a
wider stride there than the C/C++ and DXIL rules give -- and element 1 of the
buffer is then read from the wrong offset on that backend only, silently. Mesh
carried a hand-written pad for this; a pad is a derived value, so every later field
edit re-derives it, and two more structs were already divergent.

idlgen now reflects a third session on WGSL and appends a _wgslPad member to both
outputs. The padding has to be a member of the shared Slang struct: DXIL agrees
with the C++ rules, so padding either mirror alone would leave that target on the
narrower stride.

A struct's padding depends on the final size of what it embeds, so imports resolve
against the generated copies rather than the sources, and the module is reported to
Slang under its output path -- Slang looks beside the importing file before the
search paths, and would otherwise have found the unpadded source. CMake grows an
edge per import so dependencies are generated first. Without that, LoosePbrMaterial
computed its pad against an unpadded ChannelSource and came out 36 bytes wide.

Trailing padding cannot fix an interior mismatch: WGSL aligns a vector to its own
width where the scalar layout packs it tight, so PbrMaterial's float4 sat at 32 on
one side and 24 on the other. idlgen now fails naming the field, and the two
material structs lead with their widest member. Reordering is safe -- both
languages address these by name.

The stride test drops its known-divergent list; all thirteen structs now agree on
CPU, WGSL and DXIL.


* refactor(bgl/shaders): only the position matters on the degenerate path

The tail primitives of a partial meshlet were filling every interpolant by hand.
Collapsing the three corners to one point already makes the primitive zero-area,
so it rasterizes nothing and the interpolants are never read -- default-initialising
the struct and setting only the position says the same thing.

Covered by the existing test, which draws a full meshlet against a one-triangle
one: 123 of its 124 primitives take this branch, and the corners keep the clear
colour.


* refactor(bgl/idl): call the padding member pad, drop the restated comments

Review follow-up on #131.

The field-order comments on the two material structs restated a rule idlgen now
enforces: reorder them and the build fails naming the field, so the comment was
guarding nothing. MeshletInstance's comments described its two members.


---------

+++ libs/bgl/src/scene/Scene.cpp
@@ -155,8 +155,8 @@ namespace bgl
 					++trianglesDone;
 				}
 
-				meshlet.vertexCount   = static_cast<uint16_t>(localVertexCount);
-				meshlet.triangleCount = static_cast<uint16_t>(localTriangleCount);
+				meshlet.vertexCount   = localVertexCount;
+				meshlet.triangleCount = localTriangleCount;
 
 				auto minBound = glm::vec3(std::numeric_limits<float>::max());
 				auto maxBound = glm::vec3(std::numeric_limits<float>::lowest());
